requirements.txt
python-telegram-bot
requests
apscheduler
google-generativeai
python-dotenv
pytz
